Refactor test helper logic
